+++ src/main/java/com/weitaomi/application/model/bean/MemberTaskHistoryDetail.java
@@ -16,7 +16,7 @@ public class  MemberTaskHistoryDetail extends BaseModel {
      * 奖励积分
      */
     @Column(name = "pointCount")
-    private Long pointCount;
+    private Double pointCount;
 
     /**
      * 任务名称
@@ -65,7 +65,7 @@ public class  MemberTaskHistoryDetail extends BaseModel {
      *
      * @return count - 奖励积分
      */
-    public Long getPointCount() {
+    public Double getPointCount() {
         return pointCount;
     }
 
@@ -74,7 +74,7 @@ public class  MemberTaskHistoryDetail extends BaseModel {
      *
      * @param pointCount 奖励积分
      */
-    public void setPointCount(Long pointCount) {
+    public void setPointCount(Double pointCount) {
         this.pointCount = pointCount;
     }
