+++ src/main/java/com/weitaomi/application/service/impl/MemberTaskPoolService.java
@@ -67,7 +67,7 @@ public class MemberTaskPoolService extends BaseService implements IMemberTaskPoo
             }
             int num =taskPoolMapper.insertSelective(taskPool);
             memberScoreService.addMemberScore(taskPool.getMemberId(),4L,1,-Double.valueOf(taskPool.getTotalScore()), UUIDGenerator.generate());
-            boolean flag=memberTaskHistoryService.addMemberTaskToHistory(taskPool.getMemberId(),8L,-taskPool.getTotalScore().longValue(),1,"发布公众号"+officialAccount.getUserName()+"求粉任务",null);
+            boolean flag=memberTaskHistoryService.addMemberTaskToHistory(taskPool.getMemberId(),8L,-taskPool.getTotalScore().doubleValue(),1,"发布公众号"+officialAccount.getUserName()+"求粉任务",null);
             if (flag){
                 logger.info("处理成功");
                 return "任务提交成功";
@@ -148,7 +148,7 @@ public class MemberTaskPoolService extends BaseService implements IMemberTaskPoo
         taskPool.setTotalScore(taskPool.getNeedNumber()*taskPool.getSingleScore());
         int num = taskPoolMapper.insertSelective(taskPool);
         memberScoreService.addMemberScore(publishReadRequestDto.getMemberId(),5L,1,-taskPool.getNeedNumber()*taskPool.getSingleScore().doubleValue(),UUIDGenerator.generate());
-        memberTaskHistoryService.addMemberTaskToHistory(publishReadRequestDto.getMemberId(),9L,-taskPool.getNeedNumber()*taskPool.getSingleScore().longValue(),1,"发布文章\""+publishReadRequestDto.getTitle()+"\"求阅读任务",null);
+        memberTaskHistoryService.addMemberTaskToHistory(publishReadRequestDto.getMemberId(),9L,-taskPool.getNeedNumber()*taskPool.getSingleScore().doubleValue(),1,"发布文章\""+publishReadRequestDto.getTitle()+"\"求阅读任务",null);
         if (num>0) return "发布成功";
         return "发布失败";
     }
@@ -228,7 +228,7 @@ public class MemberTaskPoolService extends BaseService implements IMemberTaskPoo
             taskPool.setNeedNumber(needNumber);
             taskPool.setSingleScore(singScore);
             memberScoreService.addMemberScore(taskPool.getMemberId(),scoreId,1,-Double.valueOf(totalScore), UUIDGenerator.generate());
-            memberTaskHistoryService.addMemberTaskToHistory(taskPool.getMemberId(),taskId,-totalScore.longValue(),1,"重新发布公众号"+officialAccount.getUserName()+"求粉任务",null);
+            memberTaskHistoryService.addMemberTaskToHistory(taskPool.getMemberId(),taskId,-totalScore.doubleValue(),1,"重新发布公众号"+officialAccount.getUserName()+"求粉任务",null);
         } else if (isPublishNow==0){
             Integer score=taskPool.getTotalScore();
             taskPool.setTotalScore(0);
