+++ src/main/java/com/weitaomi/application/service/interf/IMemberTaskHistoryService.java
@@ -41,7 +41,7 @@ public interface IMemberTaskHistoryService {
      * @param
      * @return
      */
-    public boolean addMemberTaskToHistory(Long memberId, Long taskId, Long score, Integer flag,String detail,List<MemberTaskHistoryDetail> detailList);
+    public boolean addMemberTaskToHistory(Long memberId, Long taskId, Double score, Integer flag,String detail,List<MemberTaskHistoryDetail> detailList);
     /**
      * 增加任务记录
      * @param
